requirements.txt

Django>=5.0
psycopg2-binary>=2.9
python-dotenv>=1.0

djangorestframework>=3.14
djangorestframework-simplejwt>=5.3

drf-spectacular>=0.27

django-cors-headers>=4.3

django-filter>=24.3
